create new folders for Views, reorganize Login and Signup component files
